+++ tests/lib/rules/no-unknown-property.js
@@ -119,6 +119,23 @@ ruleTester.run('no-unknown-property', rule, {
     { code: '<link as="audio">Audio content</link>' },
     { code: '<video controlsList="nodownload" controls={this.controls} loop={true} muted={false} src={this.videoSrc} playsInline={true}></video>' },
     { code: '<audio controlsList="nodownload" controls={this.controls} crossOrigin="anonymous" disableRemotePlayback loop muted preload="none" src="something" onAbort={this.abort} onDurationChange={this.durationChange} onEmptied={this.emptied} onEnded={this.end} onError={this.error}></audio>' },
+    {
+      code: `
+        <table>
+          <colgroup valign="top">
+            <col valign="top" />
+          </colgroup>
+          <thead valign="top">
+            <tr valign="top">
+              <th valign="top">Header</th>
+              <td valign="top">Cell</td>
+            </tr>
+          </thead>
+          <tbody valign="top" />
+          <tfoot valign="top" />
+        </table>
+      `,
+    },
 
     // fbt
     { code: '<fbt desc="foo" doNotExtract />;' },